requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
langgraph
duckduckgo-search
ddgs
python-dotenv